+++ src/pepr/operator/controllers/istio/istio-configmap-sync.ts
@@ -6,7 +6,7 @@
 import yaml from "js-yaml";
 import { K8s, kind } from "pepr";
 import { Component, setupLogger } from "../../../logger";
-import { reloadPods } from "../secrets/reload-utils";
+import { reloadPods } from "../reload/reload-utils";
 
 export const TENANT_GATEWAY_NAMESPACE = "istio-tenant-gateway";
 export const ADMIN_GATEWAY_NAMESPACE = "istio-admin-gateway";
